requirements-agent-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic-settings==2.0.3
requests==2.31.0
langchain==0.1.16
langgraph==0.0.37
langchain-openai==0.1.3
langchain-community==0.0.32
qdrant-client==1.7.3
minio==7.1.14
tiktoken==0.6.0
# For parsing different document types.
# This installs the base library plus parsers for PDF and DOCX files.
unstructured[pdf,docx]==0.13.0
# python-magic is a dependency for filetype detection in unstructured
python-magic==0.4.27
# lxml is a dependency for some unstructured parsers
lxml==5.1.0
